Major Update: v1.0.0-alpha (#3)
# Major Update: v1.0.0-alpha
I made some breaking changes and improvements so I bumped the major version number. Specific things that were added in this PR are discussed below. I'm still testing this out and playing with it so this is just in alpha for now. 

### Swap to System.Text.Json for JSON serializer
I was using Utf8Json for JSON because it was the fastest available. However, it appears it hasn't been maintained in about a year. I did some research and it seems like System.Text.Json is the next fastest option. It also is the officially supported serializer, which I think is a positive.

### Updated packages and .NET version
I updated all the dependency packages and also upgraded to .NET 7.0. This was mostly out of a desire to stay on the latest stable stuff for security and performance purposes.

### Added a resource handler concept
I recently came across Google's [API Design Guide](https://cloud.google.com/apis/design), which talks about resource-oriented APIs and methods that map to functions, not HTTP methods. I really like this concept and it gave me a chance to fix something that always bothered me, which is that you needed to register a separate handler for "get one" vs "get all" for a collection resource. This led into creating the `ResourceHandlerBase` class.

A `ResourceHandlerBase` works just like a `RestHandlerBase` but the methods are the standard methods from Google's guide: List, Get, Create, Update, and Delete. When you register a resource handler with the service, it registers two routes: one for the collection root and one for individual items with the parameter name `id`. For example, if you register a resource handler for `/posts`, it will register routes for both `/posts` and `/posts/{id}`.